Update business-logic.rst
Encourage reusable & maintainable AND shareable ( not one big pot of "stew" aka AppBundle. I do not agree with teach them from day one to be lazy and not encapsulate functionality or let them toss in one folder mess for the next developer to break apart. Teach them early that bundles have a purpose and what that is.